requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-braces==1.13.0
django-cors-headers==3.0.2
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-guardian==2.0.0
djangorestframework==3.9.4
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3

